testing/web-platform/tests/2dcontext/drawing-images-to-the-canvas/drawimage_html_image_10_srgb.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="dest" height="100" width="100"></canvas>

<script>
var sourceWidth = 128;
var sourceHeight = 128;
var smoothingEnabled = false;
var destCanvas = document.getElementById('dest');
var sourceImg = document.createElement('img');
sourceImg.src = '../2x2-srgb.png'
sourceImg.width = sourceWidth;
sourceImg.height = sourceHeight;
var destCtx = destCanvas.getContext('2d');

function checkPixel(x, y, expected, epsilon) {
    var actual = destCtx.getImageData(x, y, 1, 1).data;
    assert_true(actual.length === expected.length);
    for (var i=0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], epsilon);
}

var pixelTests = [
    ['Pixel Top Left Image', 0, 0, [253, 140, 245, 255], 1],
    ['Pixel Top Right Image', 0, 99, [253, 140, 245, 255], 1],
    ['Pixel Bottom Left Image', 99, 0, [253, 140, 245, 255], 1],
    ['Pixel Bottom Right Image', 99, 99, [253, 140, 245, 255], 1],
];

function runTest() {
    destCtx.fillStyle = "#FF0000";
    destCtx.fillRect(0, 0,  destCanvas.width, destCanvas.height);
    destCtx.imageSmoothingEnabled = smoothingEnabled;
    // The destination rectangle is larger than the destination canvas.
    // When the destination rectangle is outside the destination image (the scratch bitmap),
    // the pixels that land outside the scratch bitmap are discarded,
    // as if the destination was an infinite canvas
    // whose rendering was clipped to the dimensions of the scratch bitmap.
    destCtx.drawImage(sourceImg, 0, 0, 512, 512, 0, 0, 256, 256);
    generate_tests(checkPixel, pixelTests);
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Draw a 100x100 image to a 100x100 canvas with resize and clipping.');
</script>
